refactor(form): remove deprecated lazy loading prefix
