+++ src/app/app.module.ts
@@ -11,7 +11,7 @@ import { AppComponent } from './app.component';
 import { MainComponent } from './main/main.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { UsersComponent } from './users/users.component';
-import { UsersFormComponent } from './users/+form/form.component';
+import { UsersFormComponent } from './users/form/form.component';
 import { LogsComponent } from './logs/logs.component';
 import { FormComponent } from './form/form.component';
 import { DetailComponent } from './detail/detail.component';
@@ -20,7 +20,7 @@ import { DashboardProductComponent } from './dashboard-product/dashboard-product
 import { ProductOverviewComponent } from './dashboard-product/overview/overview.component';
 import { ProductStatsComponent } from './dashboard-product/stats/stats.component';
 import { ProductFeaturesComponent } from './dashboard-product/features/features.component';
-import { FeaturesFormComponent } from './dashboard-product/features/+form/form.component';
+import { FeaturesFormComponent } from './dashboard-product/features/form/form.component';
 import { TemplatesComponent } from './templates/templates.component';
 import { DashboardTemplateComponent } from './templates/dashboard/dashboard.component';
 import { EmailTemplateComponent } from './templates/email/email.component';
